Catch up on the review feedback
