+++ src/ralph/configuration_management/templates/configuration_management/scminfo/scm_info.html
@@ -24,7 +24,7 @@
     </table>
 
   {% else %}
-    <p>{% trans "The status of the software configuration management is not verified yet." %}</p>
+    <p>{% trans "The status of the software configuration management for this host is unknown." %}</p>
   {% endif %}
